Changed MongoDb to NeDb (File based Database)
Brought everything in index.js
