+++ package.json
@@ -2,14 +2,14 @@
   "name": "todo-cli",
   "version": "1.0.0",
   "description": "Simple. Minimal. To-Do CLI.",
-  "main": "commands.js",
+  "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "author": "Mihir Joshi",
   "license": "ISC",
   "bin": {
-    "todo": "./commands.js"
+    "todo": "./index.js"
   },
   "dependencies": {
     "async": "^3.2.0",
@@ -19,6 +19,7 @@
     "mongodb": "^3.6.6",
     "mongojs": "^3.1.0",
     "mongoose": "^5.12.8",
+    "nedb": "^1.8.0",
     "prompts": "https://github.com/mformihir/prompts/tarball/master",
     "suppress-warnings": "^1.0.2"
   }
